move context to separate file and remove unused field props
